refactor: Modify spinner component path
