+++ src/pages/listPage/Donation/Donations.jsx
@@ -5,7 +5,7 @@ import DonationCard from './components/DonationCard';
 import FetchError from './components/FetchError';
 import { getDonations } from '@/apis/donationsApi';
 import { useApiFetch } from '@/hooks/useApiFetch';
-import Spinner from '@/components/icons/Spinner';
+import Spinner from '@/components/Spinner';
 
 export default function Donations() {
   const { data, isLoading, error } = useApiFetch(getDonations, 20);
